rename files to use conventioanal typescript names
